@@ -8,28 +8,31 @@
   ],
   "license": "MIT",
   "require": {
-    "php": "^7.2",
+    "php": "^7.4",
+    "ext-curl": "*",
     "ext-gd": "*",
+    "ext-json": "*",
     "barryvdh/laravel-ide-helper": "^2.6",
     "fideloper/proxy": "^4.0",
+    "guzzlehttp/guzzle": "^6.5",
     "intervention/image": "^2.5",
-    "laravel/framework": "^6.2",
-    "laravel/tinker": "^1.0",
-    "laravel/horizon": "^3",
-    "plancke/hypixel-php": "^1.2.0",
+    "laravel/framework": "^7",
+    "laravel/horizon": "^4",
+    "laravel/tinker": "^2.0",
+    "plancke/hypixel-php": "^1.3.0",
+    "pocketmine/nbt": "^0.2.14",
     "rcrowe/twigbridge": "^0.11.2",
     "stil/gd-text": "^1.1",
-    "ext-json": "*",
-    "ext-curl": "*"
+    "ext-zlib": "*"
   },
   "require-dev": {
     "roave/security-advisories": "dev-master",
-    "facade/ignition": "^1.4",
+    "facade/ignition": "^2.0",
     "fzaninotto/faker": "^1.4",
     "laravel/homestead": "^10",
     "mockery/mockery": "^1.0",
-    "nunomaduro/collision": "^3.0",
-    "phpunit/phpunit": "^8.0"
+    "nunomaduro/collision": "^4.1",
+    "phpunit/phpunit": "^8.5"
   },
   "config": {
     "optimize-autoloader": true,
