@@ -1,5 +1,5 @@
 <?php
-    /**
+/**
  * Copyright (c) 2020 Max Korlaar
  * All rights reserved.
  *
@@ -30,13 +30,13 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-    // @formatter:off
+// @formatter:off
 
 namespace PHPSTORM_META {
 
    /**
     * PhpStorm Meta file, to provide autocomplete information for PhpStorm
-    * Generated on 2020-05-06 17:01:07.
+    * Generated on 2020-05-17 19:32:53.
     *
     * @author Barry vd. Heuvel <barryvdh@gmail.com>
     * @see https://github.com/barryvdh/laravel-ide-helper
@@ -63,6 +63,7 @@ namespace PHPSTORM_META {
         'Illuminate\Contracts\Queue\EntityResolver' => \Illuminate\Database\Eloquent\QueueEntityResolver::class,
         'Illuminate\Contracts\Routing\ResponseFactory' => \Illuminate\Routing\ResponseFactory::class,
         'Illuminate\Foundation\Mix' => \Illuminate\Foundation\Mix::class,
+        'Illuminate\Foundation\PackageManifest' => \Illuminate\Foundation\PackageManifest::class,
         'Illuminate\Mail\Markdown' => \Illuminate\Mail\Markdown::class,
         'Illuminate\Notifications\ChannelManager' => \Illuminate\Notifications\ChannelManager::class,
         'Illuminate\Routing\Contracts\ControllerDispatcher' => \Illuminate\Routing\ControllerDispatcher::class,
@@ -82,7 +83,6 @@ namespace PHPSTORM_META {
         'Laravel\Horizon\Listeners\TrimRecentJobs' => \Laravel\Horizon\Listeners\TrimRecentJobs::class,
         'Laravel\Horizon\Lock' => \Laravel\Horizon\Lock::class,
         'Laravel\Horizon\Stopwatch' => \Laravel\Horizon\Stopwatch::class,
-        'NunoMaduro\Collision\Contracts\Adapters\Phpunit\Listener' => \NunoMaduro\Collision\Adapters\Phpunit\Listener::class,
         'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
         'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
         'auth' => \Illuminate\Auth\AuthManager::class,
@@ -96,8 +96,10 @@ namespace PHPSTORM_META {
         'command.cache.clear' => \Illuminate\Cache\Console\ClearCommand::class,
         'command.cache.forget' => \Illuminate\Cache\Console\ForgetCommand::class,
         'command.cache.table' => \Illuminate\Cache\Console\CacheTableCommand::class,
+        'command.cast.make' => \Illuminate\Foundation\Console\CastMakeCommand::class,
         'command.channel.make' => \Illuminate\Foundation\Console\ChannelMakeCommand::class,
         'command.clear-compiled' => \Illuminate\Foundation\Console\ClearCompiledCommand::class,
+        'command.component.make' => \Illuminate\Foundation\Console\ComponentMakeCommand::class,
         'command.config.cache' => \Illuminate\Foundation\Console\ConfigCacheCommand::class,
         'command.config.clear' => \Illuminate\Foundation\Console\ConfigClearCommand::class,
         'command.console.make' => \Illuminate\Foundation\Console\ConsoleMakeCommand::class,
@@ -139,7 +141,6 @@ namespace PHPSTORM_META {
         'command.optimize.clear' => \Illuminate\Foundation\Console\OptimizeClearCommand::class,
         'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
         'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
-        'command.preset' => \Illuminate\Foundation\Console\PresetCommand::class,
         'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
         'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
         'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
@@ -161,6 +162,7 @@ namespace PHPSTORM_META {
         'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
         'command.session.table' => \Illuminate\Session\Console\SessionTableCommand::class,
         'command.storage.link' => \Illuminate\Foundation\Console\StorageLinkCommand::class,
+        'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
         'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
         'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
         'command.twig' => \TwigBridge\Command\TwigBridge::class,
@@ -187,6 +189,7 @@ namespace PHPSTORM_META {
         'hash.driver' => \Illuminate\Hashing\BcryptHasher::class,
         'image' => \Intervention\Image\ImageManager::class,
         'log' => \Illuminate\Log\LogManager::class,
+        'mail.manager' => \Illuminate\Mail\MailManager::class,
         'mailer' => \Illuminate\Mail\Mailer::class,
         'memcached.connector' => \Illuminate\Cache\MemcachedConnector::class,
         'migration.creator' => \Illuminate\Database\Migrations\MigrationCreator::class,
@@ -203,8 +206,6 @@ namespace PHPSTORM_META {
         'router' => \Illuminate\Routing\Router::class,
         'session' => \Illuminate\Session\SessionManager::class,
         'session.store' => \Illuminate\Session\Store::class,
-        'swift.mailer' => \Swift_Mailer::class,
-        'swift.transport' => \Illuminate\Mail\TransportManager::class,
         'translation.loader' => \Illuminate\Translation\FileLoader::class,
         'translator' => \Illuminate\Translation\Translator::class,
         'twig' => \TwigBridge\Bridge::class,
@@ -241,6 +242,7 @@ namespace PHPSTORM_META {
         'Illuminate\Contracts\Queue\EntityResolver' => \Illuminate\Database\Eloquent\QueueEntityResolver::class,
         'Illuminate\Contracts\Routing\ResponseFactory' => \Illuminate\Routing\ResponseFactory::class,
         'Illuminate\Foundation\Mix' => \Illuminate\Foundation\Mix::class,
+        'Illuminate\Foundation\PackageManifest' => \Illuminate\Foundation\PackageManifest::class,
         'Illuminate\Mail\Markdown' => \Illuminate\Mail\Markdown::class,
         'Illuminate\Notifications\ChannelManager' => \Illuminate\Notifications\ChannelManager::class,
         'Illuminate\Routing\Contracts\ControllerDispatcher' => \Illuminate\Routing\ControllerDispatcher::class,
@@ -260,7 +262,6 @@ namespace PHPSTORM_META {
         'Laravel\Horizon\Listeners\TrimRecentJobs' => \Laravel\Horizon\Listeners\TrimRecentJobs::class,
         'Laravel\Horizon\Lock' => \Laravel\Horizon\Lock::class,
         'Laravel\Horizon\Stopwatch' => \Laravel\Horizon\Stopwatch::class,
-        'NunoMaduro\Collision\Contracts\Adapters\Phpunit\Listener' => \NunoMaduro\Collision\Adapters\Phpunit\Listener::class,
         'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
         'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
         'auth' => \Illuminate\Auth\AuthManager::class,
@@ -274,8 +275,10 @@ namespace PHPSTORM_META {
         'command.cache.clear' => \Illuminate\Cache\Console\ClearCommand::class,
         'command.cache.forget' => \Illuminate\Cache\Console\ForgetCommand::class,
         'command.cache.table' => \Illuminate\Cache\Console\CacheTableCommand::class,
+        'command.cast.make' => \Illuminate\Foundation\Console\CastMakeCommand::class,
         'command.channel.make' => \Illuminate\Foundation\Console\ChannelMakeCommand::class,
         'command.clear-compiled' => \Illuminate\Foundation\Console\ClearCompiledCommand::class,
+        'command.component.make' => \Illuminate\Foundation\Console\ComponentMakeCommand::class,
         'command.config.cache' => \Illuminate\Foundation\Console\ConfigCacheCommand::class,
         'command.config.clear' => \Illuminate\Foundation\Console\ConfigClearCommand::class,
         'command.console.make' => \Illuminate\Foundation\Console\ConsoleMakeCommand::class,
@@ -317,7 +320,6 @@ namespace PHPSTORM_META {
         'command.optimize.clear' => \Illuminate\Foundation\Console\OptimizeClearCommand::class,
         'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
         'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
-        'command.preset' => \Illuminate\Foundation\Console\PresetCommand::class,
         'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
         'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
         'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
@@ -339,6 +341,7 @@ namespace PHPSTORM_META {
         'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
         'command.session.table' => \Illuminate\Session\Console\SessionTableCommand::class,
         'command.storage.link' => \Illuminate\Foundation\Console\StorageLinkCommand::class,
+        'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
         'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
         'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
         'command.twig' => \TwigBridge\Command\TwigBridge::class,
@@ -365,6 +368,7 @@ namespace PHPSTORM_META {
         'hash.driver' => \Illuminate\Hashing\BcryptHasher::class,
         'image' => \Intervention\Image\ImageManager::class,
         'log' => \Illuminate\Log\LogManager::class,
+        'mail.manager' => \Illuminate\Mail\MailManager::class,
         'mailer' => \Illuminate\Mail\Mailer::class,
         'memcached.connector' => \Illuminate\Cache\MemcachedConnector::class,
         'migration.creator' => \Illuminate\Database\Migrations\MigrationCreator::class,
@@ -381,8 +385,6 @@ namespace PHPSTORM_META {
         'router' => \Illuminate\Routing\Router::class,
         'session' => \Illuminate\Session\SessionManager::class,
         'session.store' => \Illuminate\Session\Store::class,
-        'swift.mailer' => \Swift_Mailer::class,
-        'swift.transport' => \Illuminate\Mail\TransportManager::class,
         'translation.loader' => \Illuminate\Translation\FileLoader::class,
         'translator' => \Illuminate\Translation\Translator::class,
         'twig' => \TwigBridge\Bridge::class,
@@ -419,6 +421,7 @@ namespace PHPSTORM_META {
         'Illuminate\Contracts\Queue\EntityResolver' => \Illuminate\Database\Eloquent\QueueEntityResolver::class,
         'Illuminate\Contracts\Routing\ResponseFactory' => \Illuminate\Routing\ResponseFactory::class,
         'Illuminate\Foundation\Mix' => \Illuminate\Foundation\Mix::class,
+        'Illuminate\Foundation\PackageManifest' => \Illuminate\Foundation\PackageManifest::class,
         'Illuminate\Mail\Markdown' => \Illuminate\Mail\Markdown::class,
         'Illuminate\Notifications\ChannelManager' => \Illuminate\Notifications\ChannelManager::class,
         'Illuminate\Routing\Contracts\ControllerDispatcher' => \Illuminate\Routing\ControllerDispatcher::class,
@@ -438,7 +441,6 @@ namespace PHPSTORM_META {
         'Laravel\Horizon\Listeners\TrimRecentJobs' => \Laravel\Horizon\Listeners\TrimRecentJobs::class,
         'Laravel\Horizon\Lock' => \Laravel\Horizon\Lock::class,
         'Laravel\Horizon\Stopwatch' => \Laravel\Horizon\Stopwatch::class,
-        'NunoMaduro\Collision\Contracts\Adapters\Phpunit\Listener' => \NunoMaduro\Collision\Adapters\Phpunit\Listener::class,
         'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
         'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
         'auth' => \Illuminate\Auth\AuthManager::class,
@@ -452,8 +454,10 @@ namespace PHPSTORM_META {
         'command.cache.clear' => \Illuminate\Cache\Console\ClearCommand::class,
         'command.cache.forget' => \Illuminate\Cache\Console\ForgetCommand::class,
         'command.cache.table' => \Illuminate\Cache\Console\CacheTableCommand::class,
+        'command.cast.make' => \Illuminate\Foundation\Console\CastMakeCommand::class,
         'command.channel.make' => \Illuminate\Foundation\Console\ChannelMakeCommand::class,
         'command.clear-compiled' => \Illuminate\Foundation\Console\ClearCompiledCommand::class,
+        'command.component.make' => \Illuminate\Foundation\Console\ComponentMakeCommand::class,
         'command.config.cache' => \Illuminate\Foundation\Console\ConfigCacheCommand::class,
         'command.config.clear' => \Illuminate\Foundation\Console\ConfigClearCommand::class,
         'command.console.make' => \Illuminate\Foundation\Console\ConsoleMakeCommand::class,
@@ -495,7 +499,6 @@ namespace PHPSTORM_META {
         'command.optimize.clear' => \Illuminate\Foundation\Console\OptimizeClearCommand::class,
         'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
         'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
-        'command.preset' => \Illuminate\Foundation\Console\PresetCommand::class,
         'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
         'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
         'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
@@ -517,6 +520,7 @@ namespace PHPSTORM_META {
         'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
         'command.session.table' => \Illuminate\Session\Console\SessionTableCommand::class,
         'command.storage.link' => \Illuminate\Foundation\Console\StorageLinkCommand::class,
+        'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
         'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
         'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
         'command.twig' => \TwigBridge\Command\TwigBridge::class,
@@ -543,6 +547,7 @@ namespace PHPSTORM_META {
         'hash.driver' => \Illuminate\Hashing\BcryptHasher::class,
         'image' => \Intervention\Image\ImageManager::class,
         'log' => \Illuminate\Log\LogManager::class,
+        'mail.manager' => \Illuminate\Mail\MailManager::class,
         'mailer' => \Illuminate\Mail\Mailer::class,
         'memcached.connector' => \Illuminate\Cache\MemcachedConnector::class,
         'migration.creator' => \Illuminate\Database\Migrations\MigrationCreator::class,
@@ -559,8 +564,6 @@ namespace PHPSTORM_META {
         'router' => \Illuminate\Routing\Router::class,
         'session' => \Illuminate\Session\SessionManager::class,
         'session.store' => \Illuminate\Session\Store::class,
-        'swift.mailer' => \Swift_Mailer::class,
-        'swift.transport' => \Illuminate\Mail\TransportManager::class,
         'translation.loader' => \Illuminate\Translation\FileLoader::class,
         'translator' => \Illuminate\Translation\Translator::class,
         'twig' => \TwigBridge\Bridge::class,
@@ -597,6 +600,7 @@ namespace PHPSTORM_META {
         'Illuminate\Contracts\Queue\EntityResolver' => \Illuminate\Database\Eloquent\QueueEntityResolver::class,
         'Illuminate\Contracts\Routing\ResponseFactory' => \Illuminate\Routing\ResponseFactory::class,
         'Illuminate\Foundation\Mix' => \Illuminate\Foundation\Mix::class,
+        'Illuminate\Foundation\PackageManifest' => \Illuminate\Foundation\PackageManifest::class,
         'Illuminate\Mail\Markdown' => \Illuminate\Mail\Markdown::class,
         'Illuminate\Notifications\ChannelManager' => \Illuminate\Notifications\ChannelManager::class,
         'Illuminate\Routing\Contracts\ControllerDispatcher' => \Illuminate\Routing\ControllerDispatcher::class,
@@ -616,7 +620,6 @@ namespace PHPSTORM_META {
         'Laravel\Horizon\Listeners\TrimRecentJobs' => \Laravel\Horizon\Listeners\TrimRecentJobs::class,
         'Laravel\Horizon\Lock' => \Laravel\Horizon\Lock::class,
         'Laravel\Horizon\Stopwatch' => \Laravel\Horizon\Stopwatch::class,
-        'NunoMaduro\Collision\Contracts\Adapters\Phpunit\Listener' => \NunoMaduro\Collision\Adapters\Phpunit\Listener::class,
         'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
         'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
         'auth' => \Illuminate\Auth\AuthManager::class,
@@ -630,8 +633,10 @@ namespace PHPSTORM_META {
         'command.cache.clear' => \Illuminate\Cache\Console\ClearCommand::class,
         'command.cache.forget' => \Illuminate\Cache\Console\ForgetCommand::class,
         'command.cache.table' => \Illuminate\Cache\Console\CacheTableCommand::class,
+        'command.cast.make' => \Illuminate\Foundation\Console\CastMakeCommand::class,
         'command.channel.make' => \Illuminate\Foundation\Console\ChannelMakeCommand::class,
         'command.clear-compiled' => \Illuminate\Foundation\Console\ClearCompiledCommand::class,
+        'command.component.make' => \Illuminate\Foundation\Console\ComponentMakeCommand::class,
         'command.config.cache' => \Illuminate\Foundation\Console\ConfigCacheCommand::class,
         'command.config.clear' => \Illuminate\Foundation\Console\ConfigClearCommand::class,
         'command.console.make' => \Illuminate\Foundation\Console\ConsoleMakeCommand::class,
@@ -673,7 +678,6 @@ namespace PHPSTORM_META {
         'command.optimize.clear' => \Illuminate\Foundation\Console\OptimizeClearCommand::class,
         'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
         'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
-        'command.preset' => \Illuminate\Foundation\Console\PresetCommand::class,
         'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
         'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
         'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
@@ -695,6 +699,7 @@ namespace PHPSTORM_META {
         'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
         'command.session.table' => \Illuminate\Session\Console\SessionTableCommand::class,
         'command.storage.link' => \Illuminate\Foundation\Console\StorageLinkCommand::class,
+        'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
         'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
         'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
         'command.twig' => \TwigBridge\Command\TwigBridge::class,
@@ -721,6 +726,7 @@ namespace PHPSTORM_META {
         'hash.driver' => \Illuminate\Hashing\BcryptHasher::class,
         'image' => \Intervention\Image\ImageManager::class,
         'log' => \Illuminate\Log\LogManager::class,
+        'mail.manager' => \Illuminate\Mail\MailManager::class,
         'mailer' => \Illuminate\Mail\Mailer::class,
         'memcached.connector' => \Illuminate\Cache\MemcachedConnector::class,
         'migration.creator' => \Illuminate\Database\Migrations\MigrationCreator::class,
@@ -737,8 +743,6 @@ namespace PHPSTORM_META {
         'router' => \Illuminate\Routing\Router::class,
         'session' => \Illuminate\Session\SessionManager::class,
         'session.store' => \Illuminate\Session\Store::class,
-        'swift.mailer' => \Swift_Mailer::class,
-        'swift.transport' => \Illuminate\Mail\TransportManager::class,
         'translation.loader' => \Illuminate\Translation\FileLoader::class,
         'translator' => \Illuminate\Translation\Translator::class,
         'twig' => \TwigBridge\Bridge::class,
@@ -775,6 +779,7 @@ namespace PHPSTORM_META {
         'Illuminate\Contracts\Queue\EntityResolver' => \Illuminate\Database\Eloquent\QueueEntityResolver::class,
         'Illuminate\Contracts\Routing\ResponseFactory' => \Illuminate\Routing\ResponseFactory::class,
         'Illuminate\Foundation\Mix' => \Illuminate\Foundation\Mix::class,
+        'Illuminate\Foundation\PackageManifest' => \Illuminate\Foundation\PackageManifest::class,
         'Illuminate\Mail\Markdown' => \Illuminate\Mail\Markdown::class,
         'Illuminate\Notifications\ChannelManager' => \Illuminate\Notifications\ChannelManager::class,
         'Illuminate\Routing\Contracts\ControllerDispatcher' => \Illuminate\Routing\ControllerDispatcher::class,
@@ -794,7 +799,6 @@ namespace PHPSTORM_META {
         'Laravel\Horizon\Listeners\TrimRecentJobs' => \Laravel\Horizon\Listeners\TrimRecentJobs::class,
         'Laravel\Horizon\Lock' => \Laravel\Horizon\Lock::class,
         'Laravel\Horizon\Stopwatch' => \Laravel\Horizon\Stopwatch::class,
-        'NunoMaduro\Collision\Contracts\Adapters\Phpunit\Listener' => \NunoMaduro\Collision\Adapters\Phpunit\Listener::class,
         'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
         'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
         'auth' => \Illuminate\Auth\AuthManager::class,
@@ -808,8 +812,10 @@ namespace PHPSTORM_META {
         'command.cache.clear' => \Illuminate\Cache\Console\ClearCommand::class,
         'command.cache.forget' => \Illuminate\Cache\Console\ForgetCommand::class,
         'command.cache.table' => \Illuminate\Cache\Console\CacheTableCommand::class,
+        'command.cast.make' => \Illuminate\Foundation\Console\CastMakeCommand::class,
         'command.channel.make' => \Illuminate\Foundation\Console\ChannelMakeCommand::class,
         'command.clear-compiled' => \Illuminate\Foundation\Console\ClearCompiledCommand::class,
+        'command.component.make' => \Illuminate\Foundation\Console\ComponentMakeCommand::class,
         'command.config.cache' => \Illuminate\Foundation\Console\ConfigCacheCommand::class,
         'command.config.clear' => \Illuminate\Foundation\Console\ConfigClearCommand::class,
         'command.console.make' => \Illuminate\Foundation\Console\ConsoleMakeCommand::class,
@@ -851,7 +857,6 @@ namespace PHPSTORM_META {
         'command.optimize.clear' => \Illuminate\Foundation\Console\OptimizeClearCommand::class,
         'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
         'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
-        'command.preset' => \Illuminate\Foundation\Console\PresetCommand::class,
         'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
         'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
         'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
@@ -873,6 +878,7 @@ namespace PHPSTORM_META {
         'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
         'command.session.table' => \Illuminate\Session\Console\SessionTableCommand::class,
         'command.storage.link' => \Illuminate\Foundation\Console\StorageLinkCommand::class,
+        'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
         'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
         'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
         'command.twig' => \TwigBridge\Command\TwigBridge::class,
@@ -899,6 +905,7 @@ namespace PHPSTORM_META {
         'hash.driver' => \Illuminate\Hashing\BcryptHasher::class,
         'image' => \Intervention\Image\ImageManager::class,
         'log' => \Illuminate\Log\LogManager::class,
+        'mail.manager' => \Illuminate\Mail\MailManager::class,
         'mailer' => \Illuminate\Mail\Mailer::class,
         'memcached.connector' => \Illuminate\Cache\MemcachedConnector::class,
         'migration.creator' => \Illuminate\Database\Migrations\MigrationCreator::class,
@@ -915,8 +922,6 @@ namespace PHPSTORM_META {
         'router' => \Illuminate\Routing\Router::class,
         'session' => \Illuminate\Session\SessionManager::class,
         'session.store' => \Illuminate\Session\Store::class,
-        'swift.mailer' => \Swift_Mailer::class,
-        'swift.transport' => \Illuminate\Mail\TransportManager::class,
         'translation.loader' => \Illuminate\Translation\FileLoader::class,
         'translator' => \Illuminate\Translation\Translator::class,
         'twig' => \TwigBridge\Bridge::class,
@@ -953,6 +958,7 @@ namespace PHPSTORM_META {
         'Illuminate\Contracts\Queue\EntityResolver' => \Illuminate\Database\Eloquent\QueueEntityResolver::class,
         'Illuminate\Contracts\Routing\ResponseFactory' => \Illuminate\Routing\ResponseFactory::class,
         'Illuminate\Foundation\Mix' => \Illuminate\Foundation\Mix::class,
+        'Illuminate\Foundation\PackageManifest' => \Illuminate\Foundation\PackageManifest::class,
         'Illuminate\Mail\Markdown' => \Illuminate\Mail\Markdown::class,
         'Illuminate\Notifications\ChannelManager' => \Illuminate\Notifications\ChannelManager::class,
         'Illuminate\Routing\Contracts\ControllerDispatcher' => \Illuminate\Routing\ControllerDispatcher::class,
@@ -972,7 +978,6 @@ namespace PHPSTORM_META {
         'Laravel\Horizon\Listeners\TrimRecentJobs' => \Laravel\Horizon\Listeners\TrimRecentJobs::class,
         'Laravel\Horizon\Lock' => \Laravel\Horizon\Lock::class,
         'Laravel\Horizon\Stopwatch' => \Laravel\Horizon\Stopwatch::class,
-        'NunoMaduro\Collision\Contracts\Adapters\Phpunit\Listener' => \NunoMaduro\Collision\Adapters\Phpunit\Listener::class,
         'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
         'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
         'auth' => \Illuminate\Auth\AuthManager::class,
@@ -986,8 +991,10 @@ namespace PHPSTORM_META {
         'command.cache.clear' => \Illuminate\Cache\Console\ClearCommand::class,
         'command.cache.forget' => \Illuminate\Cache\Console\ForgetCommand::class,
         'command.cache.table' => \Illuminate\Cache\Console\CacheTableCommand::class,
+        'command.cast.make' => \Illuminate\Foundation\Console\CastMakeCommand::class,
         'command.channel.make' => \Illuminate\Foundation\Console\ChannelMakeCommand::class,
         'command.clear-compiled' => \Illuminate\Foundation\Console\ClearCompiledCommand::class,
+        'command.component.make' => \Illuminate\Foundation\Console\ComponentMakeCommand::class,
         'command.config.cache' => \Illuminate\Foundation\Console\ConfigCacheCommand::class,
         'command.config.clear' => \Illuminate\Foundation\Console\ConfigClearCommand::class,
         'command.console.make' => \Illuminate\Foundation\Console\ConsoleMakeCommand::class,
@@ -1029,7 +1036,6 @@ namespace PHPSTORM_META {
         'command.optimize.clear' => \Illuminate\Foundation\Console\OptimizeClearCommand::class,
         'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
         'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
-        'command.preset' => \Illuminate\Foundation\Console\PresetCommand::class,
         'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
         'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
         'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
@@ -1051,6 +1057,7 @@ namespace PHPSTORM_META {
         'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
         'command.session.table' => \Illuminate\Session\Console\SessionTableCommand::class,
         'command.storage.link' => \Illuminate\Foundation\Console\StorageLinkCommand::class,
+        'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
         'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
         'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
         'command.twig' => \TwigBridge\Command\TwigBridge::class,
@@ -1077,6 +1084,7 @@ namespace PHPSTORM_META {
         'hash.driver' => \Illuminate\Hashing\BcryptHasher::class,
         'image' => \Intervention\Image\ImageManager::class,
         'log' => \Illuminate\Log\LogManager::class,
+        'mail.manager' => \Illuminate\Mail\MailManager::class,
         'mailer' => \Illuminate\Mail\Mailer::class,
         'memcached.connector' => \Illuminate\Cache\MemcachedConnector::class,
         'migration.creator' => \Illuminate\Database\Migrations\MigrationCreator::class,
@@ -1093,8 +1101,6 @@ namespace PHPSTORM_META {
         'router' => \Illuminate\Routing\Router::class,
         'session' => \Illuminate\Session\SessionManager::class,
         'session.store' => \Illuminate\Session\Store::class,
-        'swift.mailer' => \Swift_Mailer::class,
-        'swift.transport' => \Illuminate\Mail\TransportManager::class,
         'translation.loader' => \Illuminate\Translation\FileLoader::class,
         'translator' => \Illuminate\Translation\Translator::class,
         'twig' => \TwigBridge\Bridge::class,
@@ -1131,6 +1137,7 @@ namespace PHPSTORM_META {
         'Illuminate\Contracts\Queue\EntityResolver' => \Illuminate\Database\Eloquent\QueueEntityResolver::class,
         'Illuminate\Contracts\Routing\ResponseFactory' => \Illuminate\Routing\ResponseFactory::class,
         'Illuminate\Foundation\Mix' => \Illuminate\Foundation\Mix::class,
+        'Illuminate\Foundation\PackageManifest' => \Illuminate\Foundation\PackageManifest::class,
         'Illuminate\Mail\Markdown' => \Illuminate\Mail\Markdown::class,
         'Illuminate\Notifications\ChannelManager' => \Illuminate\Notifications\ChannelManager::class,
         'Illuminate\Routing\Contracts\ControllerDispatcher' => \Illuminate\Routing\ControllerDispatcher::class,
@@ -1150,7 +1157,6 @@ namespace PHPSTORM_META {
         'Laravel\Horizon\Listeners\TrimRecentJobs' => \Laravel\Horizon\Listeners\TrimRecentJobs::class,
         'Laravel\Horizon\Lock' => \Laravel\Horizon\Lock::class,
         'Laravel\Horizon\Stopwatch' => \Laravel\Horizon\Stopwatch::class,
-        'NunoMaduro\Collision\Contracts\Adapters\Phpunit\Listener' => \NunoMaduro\Collision\Adapters\Phpunit\Listener::class,
         'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
         'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
         'auth' => \Illuminate\Auth\AuthManager::class,
@@ -1164,8 +1170,10 @@ namespace PHPSTORM_META {
         'command.cache.clear' => \Illuminate\Cache\Console\ClearCommand::class,
         'command.cache.forget' => \Illuminate\Cache\Console\ForgetCommand::class,
         'command.cache.table' => \Illuminate\Cache\Console\CacheTableCommand::class,
+        'command.cast.make' => \Illuminate\Foundation\Console\CastMakeCommand::class,
         'command.channel.make' => \Illuminate\Foundation\Console\ChannelMakeCommand::class,
         'command.clear-compiled' => \Illuminate\Foundation\Console\ClearCompiledCommand::class,
+        'command.component.make' => \Illuminate\Foundation\Console\ComponentMakeCommand::class,
         'command.config.cache' => \Illuminate\Foundation\Console\ConfigCacheCommand::class,
         'command.config.clear' => \Illuminate\Foundation\Console\ConfigClearCommand::class,
         'command.console.make' => \Illuminate\Foundation\Console\ConsoleMakeCommand::class,
@@ -1207,7 +1215,6 @@ namespace PHPSTORM_META {
         'command.optimize.clear' => \Illuminate\Foundation\Console\OptimizeClearCommand::class,
         'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
         'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
-        'command.preset' => \Illuminate\Foundation\Console\PresetCommand::class,
         'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
         'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
         'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
@@ -1229,6 +1236,7 @@ namespace PHPSTORM_META {
         'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
         'command.session.table' => \Illuminate\Session\Console\SessionTableCommand::class,
         'command.storage.link' => \Illuminate\Foundation\Console\StorageLinkCommand::class,
+        'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
         'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
         'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
         'command.twig' => \TwigBridge\Command\TwigBridge::class,
@@ -1255,6 +1263,7 @@ namespace PHPSTORM_META {
         'hash.driver' => \Illuminate\Hashing\BcryptHasher::class,
         'image' => \Intervention\Image\ImageManager::class,
         'log' => \Illuminate\Log\LogManager::class,
+        'mail.manager' => \Illuminate\Mail\MailManager::class,
         'mailer' => \Illuminate\Mail\Mailer::class,
         'memcached.connector' => \Illuminate\Cache\MemcachedConnector::class,
         'migration.creator' => \Illuminate\Database\Migrations\MigrationCreator::class,
@@ -1271,8 +1280,6 @@ namespace PHPSTORM_META {
         'router' => \Illuminate\Routing\Router::class,
         'session' => \Illuminate\Session\SessionManager::class,
         'session.store' => \Illuminate\Session\Store::class,
-        'swift.mailer' => \Swift_Mailer::class,
-        'swift.transport' => \Illuminate\Mail\TransportManager::class,
         'translation.loader' => \Illuminate\Translation\FileLoader::class,
         'translator' => \Illuminate\Translation\Translator::class,
         'twig' => \TwigBridge\Bridge::class,
@@ -1309,6 +1316,7 @@ namespace PHPSTORM_META {
         'Illuminate\Contracts\Queue\EntityResolver' => \Illuminate\Database\Eloquent\QueueEntityResolver::class,
         'Illuminate\Contracts\Routing\ResponseFactory' => \Illuminate\Routing\ResponseFactory::class,
         'Illuminate\Foundation\Mix' => \Illuminate\Foundation\Mix::class,
+        'Illuminate\Foundation\PackageManifest' => \Illuminate\Foundation\PackageManifest::class,
         'Illuminate\Mail\Markdown' => \Illuminate\Mail\Markdown::class,
         'Illuminate\Notifications\ChannelManager' => \Illuminate\Notifications\ChannelManager::class,
         'Illuminate\Routing\Contracts\ControllerDispatcher' => \Illuminate\Routing\ControllerDispatcher::class,
@@ -1328,7 +1336,6 @@ namespace PHPSTORM_META {
         'Laravel\Horizon\Listeners\TrimRecentJobs' => \Laravel\Horizon\Listeners\TrimRecentJobs::class,
         'Laravel\Horizon\Lock' => \Laravel\Horizon\Lock::class,
         'Laravel\Horizon\Stopwatch' => \Laravel\Horizon\Stopwatch::class,
-        'NunoMaduro\Collision\Contracts\Adapters\Phpunit\Listener' => \NunoMaduro\Collision\Adapters\Phpunit\Listener::class,
         'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
         'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
         'auth' => \Illuminate\Auth\AuthManager::class,
@@ -1342,8 +1349,10 @@ namespace PHPSTORM_META {
         'command.cache.clear' => \Illuminate\Cache\Console\ClearCommand::class,
         'command.cache.forget' => \Illuminate\Cache\Console\ForgetCommand::class,
         'command.cache.table' => \Illuminate\Cache\Console\CacheTableCommand::class,
+        'command.cast.make' => \Illuminate\Foundation\Console\CastMakeCommand::class,
         'command.channel.make' => \Illuminate\Foundation\Console\ChannelMakeCommand::class,
         'command.clear-compiled' => \Illuminate\Foundation\Console\ClearCompiledCommand::class,
+        'command.component.make' => \Illuminate\Foundation\Console\ComponentMakeCommand::class,
         'command.config.cache' => \Illuminate\Foundation\Console\ConfigCacheCommand::class,
         'command.config.clear' => \Illuminate\Foundation\Console\ConfigClearCommand::class,
         'command.console.make' => \Illuminate\Foundation\Console\ConsoleMakeCommand::class,
@@ -1385,7 +1394,6 @@ namespace PHPSTORM_META {
         'command.optimize.clear' => \Illuminate\Foundation\Console\OptimizeClearCommand::class,
         'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
         'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
-        'command.preset' => \Illuminate\Foundation\Console\PresetCommand::class,
         'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
         'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
         'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
@@ -1407,6 +1415,7 @@ namespace PHPSTORM_META {
         'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
         'command.session.table' => \Illuminate\Session\Console\SessionTableCommand::class,
         'command.storage.link' => \Illuminate\Foundation\Console\StorageLinkCommand::class,
+        'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
         'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
         'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
         'command.twig' => \TwigBridge\Command\TwigBridge::class,
@@ -1433,6 +1442,7 @@ namespace PHPSTORM_META {
         'hash.driver' => \Illuminate\Hashing\BcryptHasher::class,
         'image' => \Intervention\Image\ImageManager::class,
         'log' => \Illuminate\Log\LogManager::class,
+        'mail.manager' => \Illuminate\Mail\MailManager::class,
         'mailer' => \Illuminate\Mail\Mailer::class,
         'memcached.connector' => \Illuminate\Cache\MemcachedConnector::class,
         'migration.creator' => \Illuminate\Database\Migrations\MigrationCreator::class,
@@ -1449,8 +1459,6 @@ namespace PHPSTORM_META {
         'router' => \Illuminate\Routing\Router::class,
         'session' => \Illuminate\Session\SessionManager::class,
         'session.store' => \Illuminate\Session\Store::class,
-        'swift.mailer' => \Swift_Mailer::class,
-        'swift.transport' => \Illuminate\Mail\TransportManager::class,
         'translation.loader' => \Illuminate\Translation\FileLoader::class,
         'translator' => \Illuminate\Translation\Translator::class,
         'twig' => \TwigBridge\Bridge::class,
