@@ -36,6 +36,7 @@
     use Illuminate\Foundation\Exceptions\Handler as ExceptionHandler;
     use Illuminate\Http\Request;
     use Illuminate\Http\Response;
+    use Throwable;
 
     /**
      * Class Handler
@@ -65,12 +66,12 @@
         /**
          * Report or log an exception.
          *
-         * @param Exception $exception
+         * @param Throwable $exception
          *
          * @return void
          * @throws Exception
          */
-        public function report(Exception $exception) {
+        public function report(Throwable $exception) {
             parent::report($exception);
         }
 
@@ -78,11 +79,12 @@
          * Render an exception into an HTTP response.
          *
          * @param Request   $request
-         * @param Exception $exception
+         * @param Throwable $exception
          *
          * @return Response
+         * @throws Throwable
          */
-        public function render($request, Exception $exception) {
+        public function render($request, Throwable $exception) {
             return parent::render($request, $exception);
         }
     }
