@@ -126,7 +126,10 @@
 
             $this->addWatermark($image, $fontSourceSansProLight, 650, 160); // Watermark/advertisement
 
-            return Image::make($image)->response('png');
+            return Image::make($image)->response('png')->setCache([
+                'public'  => true,
+                'max_age' => 600
+            ]);
         }
 
     }
