@@ -1,5 +1,5 @@
 <?php
-/**
+    /**
  * Copyright (c) 2020 Max Korlaar
  * All rights reserved.
  *
@@ -54,7 +54,7 @@
             $this->sessionURL = 'https://sessionserver.mojang.com/session/minecraft/profile/';
             $this->profileURL = 'https://api.mojang.com/users/profiles/minecraft/';
             $this->timeout    = config('signatures.api_timeout');
-            $this->cacheTime  = 2 * 7200;
+            $this->cacheTime  = config('cache.times.mojang_api');
         }
 
         /**
@@ -73,7 +73,13 @@
             $request = $this->request($this->sessionURL . $uuid);
 
             if ($request['success'] === false) {
-                Log::debug('Something went wrong while fetching profile data', [$request]);
+                if ($this->checkForThrottle($request)) {
+                    Log::debug('Could not fetch profile from Mojang API due to throttle', ['uuid' => $uuid]);
+
+                    return ['success' => false, 'throttle' => true];
+                }
+
+                Log::warning('Something went wrong while fetching profile data', ['url' => $this->sessionURL . $uuid, 'return' => $request]);
                 return $request;
             }
 
@@ -101,6 +107,8 @@
             }
             $profileData = ['skinURL' => $skinURL, 'isSteve' => $isSteve, 'username' => $jsonArray['name'], 'uuid' => $jsonArray['id']];
             $return      = ['success' => true, 'data' => $profileData];
+
+            Log::debug('Fetched profile from Mojang', ['uuid' => $uuid, 'data' => $profileData]);
 
             Cache::set($cacheKey, $return, $this->getCacheTime());
 
@@ -134,27 +142,17 @@
             if ($status !== 200) {
                 return ['success' => false, 'status_code' => $status, 'error' => null];
             }
+
             return ['success' => true, 'data' => $curlOut];
         }
 
         /**
-         * @param $username
+         * @param $request
          *
-         * @return array
+         * @return bool
          */
-        public function getUUID($username): array {
-            return Cache::remember('mojangapi.username.' . $username, $this->getCacheTime(), function () use ($username) {
-                $request = $this->request($this->profileURL . $username);
-                if ($request['success'] === false) {
-                    if ($this->checkForThrottle($request)) {
-                        return ['success' => false, 'throttle' => true];
-                    }
-                    return $request;
-                }
-                $jsonArray = json_decode($request['data'], true);
-                return ['success' => true, 'data' => $jsonArray];
-            });
-
+        public function checkForThrottle($request): bool {
+            return isset($request['status_code']) && $request['status_code'] === 429;
         }
 
         /**
@@ -172,12 +170,37 @@
         }
 
         /**
-         * @param $request
+         * @param $username
          *
-         * @return bool
+         * @return array
+         * @throws InvalidArgumentException
          */
-        public function checkForThrottle($request): bool {
-            return isset($request['status_code']) && $request['status_code'] === 429;
+        public function getUUID($username): array {
+            $cacheKey = 'mojangapi.username.' . $username;
+
+            if (Cache::has($cacheKey)) {
+                return Cache::get($cacheKey);
+            }
+
+            $request = $this->request($this->profileURL . $username);
+
+            if ($request['success'] === false) {
+                if ($this->checkForThrottle($request)) {
+                    Log::debug('Could not fetch UUID from Mojang API due to throttle', ['username' => $username]);
+
+                    return ['success' => false, 'throttle' => true];
+                }
+                return $request;
+            }
+
+            $jsonArray = json_decode($request['data'], true);
+            $return    = ['success' => true, 'data' => $jsonArray];
+
+            Log::debug('Fetched UUID from Mojang API', ['username' => $username, 'data' => $jsonArray]);
+
+            Cache::set($cacheKey, $return, $this->getCacheTime());
+
+            return $return;
         }
 
         /**
