@@ -1,5 +1,5 @@
 <?php
-/**
+    /**
  * Copyright (c) 2020 Max Korlaar
  * All rights reserved.
  *
@@ -85,7 +85,9 @@
         'url'         => env('APP_URL', 'http://localhost'),
         'force_https' => env('FORCE_HTTPS', true),
 
+        'signature_debug_background'   => env('SIGNATURE_DEBUG_BACKGROUND', false),
         'google_analytics_tracking_id' => 'UA-157532904-4',
+        'google_ads_client_id'         => 'ca-pub-9570587520778628',
 
         'asset_url' => env('ASSET_URL', null),
 
@@ -155,7 +157,6 @@
         'key' => env('APP_KEY'),
 
         'cipher' => 'AES-256-CBC',
-
 
         /*
         |--------------------------------------------------------------------------
