@@ -30,49 +30,76 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-    return [
+return [
 
         /*
         |--------------------------------------------------------------------------
-        | Mail Driver
+        | Default Mailer
         |--------------------------------------------------------------------------
         |
-        | Laravel supports both SMTP and PHP's "mail" function as drivers for the
-        | sending of e-mail. You may specify which one you're using throughout
-        | your application here. By default, Laravel is setup for SMTP mail.
+        | This option controls the default mailer that is used to send any email
+        | messages sent by your application. Alternative mailers may be setup
+        | and used as needed; however, this mailer will be used by default.
+        |
+        */
+
+        'default' => env('MAIL_MAILER', 'smtp'),
+
+        /*
+        |--------------------------------------------------------------------------
+        | Mailer Configurations
+        |--------------------------------------------------------------------------
+        |
+        | Here you may configure all of the mailers used by your application plus
+        | their respective settings. Several examples have been configured for
+        | you and you are free to add your own as your application requires.
+        |
+        | Laravel supports a variety of mail "transport" drivers to be used while
+        | sending an e-mail. You will specify which one you are using for your
+        | mailers below. You are free to add additional mailers as required.
         |
         | Supported: "smtp", "sendmail", "mailgun", "ses",
         |            "postmark", "log", "array"
         |
         */
 
-        'driver' => env('MAIL_DRIVER', 'smtp'),
+        'mailers' => [
+            'smtp' => [
+                'transport'  => 'smtp',
+                'host'       => env('MAIL_HOST', 'smtp.mailgun.org'),
+                'port'       => env('MAIL_PORT', 587),
+                'encryption' => env('MAIL_ENCRYPTION', 'tls'),
+                'username'   => env('MAIL_USERNAME'),
+                'password'   => env('MAIL_PASSWORD'),
+                'timeout'    => null,
+            ],
 
-        /*
-        |--------------------------------------------------------------------------
-        | SMTP Host Address
-        |--------------------------------------------------------------------------
-        |
-        | Here you may provide the host address of the SMTP server used by your
-        | applications. A default option is provided that is compatible with
-        | the Mailgun mail service which will provide reliable deliveries.
-        |
-        */
+            'ses' => [
+                'transport' => 'ses',
+            ],
 
-        'host' => env('MAIL_HOST', 'smtp.mailgun.org'),
+            'mailgun' => [
+                'transport' => 'mailgun',
+            ],
 
-        /*
-        |--------------------------------------------------------------------------
-        | SMTP Host Port
-        |--------------------------------------------------------------------------
-        |
-        | This is the SMTP port used by your application to deliver e-mails to
-        | users of the application. Like the host we have set this value to
-        | stay compatible with the Mailgun e-mail application by default.
-        |
-        */
+            'postmark' => [
+                'transport' => 'postmark',
+            ],
 
-        'port' => env('MAIL_PORT', 587),
+            'sendmail' => [
+                'transport' => 'sendmail',
+                'path'      => '/usr/sbin/sendmail -bs',
+            ],
+
+            'log' => [
+                'transport' => 'log',
+                'channel'   => env('MAIL_LOG_CHANNEL'),
+            ],
+
+            'array' => [
+                'transport' => 'array',
+            ],
+        ],
 
         /*
         |--------------------------------------------------------------------------
@@ -92,47 +119,6 @@
 
         /*
         |--------------------------------------------------------------------------
-        | E-Mail Encryption Protocol
-        |--------------------------------------------------------------------------
-        |
-        | Here you may specify the encryption protocol that should be used when
-        | the application send e-mail messages. A sensible default using the
-        | transport layer security protocol should provide great security.
-        |
-        */
-
-        'encryption' => env('MAIL_ENCRYPTION', 'tls'),
-
-        /*
-        |--------------------------------------------------------------------------
-        | SMTP Server Username
-        |--------------------------------------------------------------------------
-        |
-        | If your SMTP server requires a username for authentication, you should
-        | set it here. This will get used to authenticate with your server on
-        | connection. You may also set the "password" value below this one.
-        |
-        */
-
-        'username' => env('MAIL_USERNAME'),
-
-        'password' => env('MAIL_PASSWORD'),
-
-        /*
-        |--------------------------------------------------------------------------
-        | Sendmail System Path
-        |--------------------------------------------------------------------------
-        |
-        | When using the "sendmail" driver to send e-mails, we will need to know
-        | the path to where Sendmail lives on this server. A default path has
-        | been provided here, which will work well on most of your systems.
-        |
-        */
-
-        'sendmail' => '/usr/sbin/sendmail -bs',
-
-        /*
-        |--------------------------------------------------------------------------
         | Markdown Mail Settings
         |--------------------------------------------------------------------------
         |
@@ -149,18 +135,5 @@
                 resource_path('views/vendor/mail'),
             ],
         ],
-
-        /*
-        |--------------------------------------------------------------------------
-        | Log Channel
-        |--------------------------------------------------------------------------
-        |
-        | If you are using the "log" driver, you may specify the logging channel
-        | if you prefer to keep mail messages separate from other log entries
-        | for simpler reading. Otherwise, the default channel will be used.
-        |
-        */
-
-        'log_channel' => env('MAIL_LOG_CHANNEL'),
 
     ];
