@@ -196,7 +196,7 @@
         |
         */
 
-        'secure' => env('SESSION_SECURE_COOKIE', false),
+        'secure' => env('SESSION_SECURE_COOKIE', null),
 
         /*
         |--------------------------------------------------------------------------
