@@ -45,7 +45,7 @@
          *
          * @return void
          */
-        public function testBasicTest() {
+        public function testBasicTest(): void {
             $response = $this->get('/');
 
             $response->assertStatus(200);
